==> 3 - Visualization and analyses.Rmd <==
---
title: "3 - Visualization"
author: "Álvaro Sanz"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(ggplot2)
library(data.table)
library(lubridate)
library(showtext)
```


# Importing the data

```{r}
#dfull <- read.csv("data/predicted.csv")
dfull <- fread("data/predicted.csv")
```

# Sample of the predicted tweets

```{r}
table(dfull$hate)
```

```{r}
set.seed(2531)
num_tweets_to_display <- 3
hateful_tweets <- dfull[dfull$hate == "Yes", ]
random_tweets <- hateful_tweets[sample(nrow(hateful_tweets), num_tweets_to_display), ]
table <- random_tweets[,text]
table
```

# Data transforming

```{r}
min(dfull$created_at)
max(dfull$created_at)
```

```{r}
dfull2 <- dfull
dfull <- dfull %>% 
  mutate(created_at = ymd_hms(created_at),
         created_at = format(created_at, "%Y-%m-%d %H:00:00"))

viz <- dfull %>% 
  mutate(hate = factor(hate, levels = c("No", "Yes"))) %>% 
  group_by(created_at) %>% 
  summarise(No_count = sum(hate == "No", na.rm = TRUE),
            Yes_count = sum(hate == "Yes", na.rm = TRUE),
            Total = sum(No_count, Yes_count),
            Proportion = Yes_count/Total * 100)

```

```{r}
viz <- viz %>% 
  mutate(match = case_when(created_at == "2022-11-23 15:00:00" | created_at == "2022-11-23 16:00:00" | created_at == "2022-11-23 17:00:00" | created_at == "2022-11-23 18:00:00" | created_at == "2022-11-23 19:00:00" | created_at == "2022-11-23 20:00:00" | created_at == "2022-11-23 21:00:00" | created_at == "2022-11-27 17:00:00" | created_at == "2022-11-27 18:00:00" | created_at == "2022-11-27 19:00:00" | created_at == "2022-11-27 20:00:00" | created_at == "2022-11-27 21:00:00" | created_at == "2022-11-27 22:00:00" | created_at == "2022-11-27 23:00:00" | created_at == "2022-11-28 00:00:00" | created_at == "2022-12-01 18:00:00" | created_at == "2022-12-01 19:00:00" | created_at == "2022-12-01 20:00:00" | created_at == "2022-12-01 21:00:00" | created_at == "2022-12-01 22:00:00" | created_at == "2022-12-01 23:00:00" | created_at == "2022-12-02 00:00:00" | created_at == "2022-12-06 14:00:00" | created_at == "2022-12-06 15:00:00" | created_at == "2022-12-06 16:00:00" | created_at == "2022-12-06 17:00:00" | created_at == "2022-12-06 18:00:00" | created_at == "2022-12-06 19:00:00" ~ 1,
                           T ~ 0),
         match = factor(match, levels = c(0,1)), 
         costarica = factor(ifelse((match == 1 & day(created_at) == 23), 1, 0), levels = c(0,1)),
         germany = factor(ifelse((match == 1 & day(created_at) == 27), 1, 0), levels = c(0,1)),
         japan = factor(ifelse((match == 1 & day(created_at) == 01), 1, 0), levels = c(0,1)),
         morocco = factor(ifelse((match == 1 & day(created_at) == 06), 1, 0), levels = c(0,1)))
```

# Variance checking: 

```{r}
varfull <- var(viz$Proportion)

varcheck <- viz %>% 
  filter(created_at < "2022-12-05 00:00:00")

varpre <- var(varcheck$Proportion)

variances <- as.data.frame(table(varfull, varpre)) %>% 
  select(-Freq) %>% 
  rename("Variance for full period" = "varfull",
         "Variance for first 3 games" = "varpre") %>% 
  mutate_all(~ as.numeric(as.character(.))) %>%
  mutate(across(everything(), ~ round(., 2)))

kableExtra::kable(variances, format = "pipe", align = "c")
```


# Visualization

```{r}
sysfonts::font_add_google("Titillium Web", family="titillium")
showtext::showtext_auto()
```

```{r}
library(png)
germanyflag <- readPNG("data/images/1200px-Flag_of_Germany.png")
japanflag <- readPNG("data/images/Flag_of_Japan.png")
crflag <- readPNG("data/images/Flag_of_Costa_Rica.png")
morflag <- readPNG("data/images/Flag_of_Morocco.png")
```


```{r,fig.height=5, fig.width=8,  out.height="100%", out.width="100%"}
library(grid)

vizjap <- viz %>% 
  filter(match == 1 & japan == 1)
vizcr <- viz %>% 
  filter(match == 1 & costarica == 1)
vizger <- viz %>% 
  filter(match == 1 & germany == 1)
vizmor <- viz %>% 
  filter(match == 1 & morocco == 1)



viz %>% 
  ggplot() +
  aes(x = as.POSIXct(created_at), y = Proportion) +
  geom_rect(aes(xmin=as.POSIXct("2022-11-23 15:00"), xmax=as.POSIXct("2022-11-23 21:00"), ymin=(-Inf), ymax=Inf),
            alpha=0.1, show.legend=FALSE, fill = "lightgrey") + #Costa Rica
  geom_rect(aes(xmin=as.POSIXct("2022-11-27 18:00"), xmax=as.POSIXct("2022-11-28 00:00"), ymin=(-Inf), ymax=Inf),
            alpha=0.1, show.legend=FALSE, fill = "lightgrey") + #Germany
  geom_rect(aes(xmin=as.POSIXct("2022-12-01 18:00"), xmax=as.POSIXct("2022-12-02 00:00"), ymin=(-Inf), ymax=Inf),
            alpha=0.1, show.legend=FALSE, fill = "lightgrey") + #Japan
  geom_rect(aes(xmin=as.POSIXct("2022-12-06 14:00"), xmax=as.POSIXct("2022-12-06 19:00"), ymin=(-Inf), ymax=Inf),
            alpha=0.1, show.legend=FALSE, fill = "lightgrey") + #Morocco
  geom_line(size = 0.3) +
  geom_smooth(method = "loess", size = 0.3) +
  geom_line(data = vizjap, aes(x = as.POSIXct(created_at), y = Proportion), color = "orange") +
  geom_line(data = vizcr, aes(x = as.POSIXct(created_at), y = Proportion), color = "darkgreen") +
  geom_line(data = vizger, aes(x = as.POSIXct(created_at), y = Proportion), color = "orange") +
  geom_line(data = vizmor, aes(x = as.POSIXct(created_at), y = Proportion), color = "red") +
  scale_color_manual(aes(match), values = c("black", "orange")) +
  scale_y_continuous(
    limits = c(0,6),
    labels = function(x) paste0(x, "%")) +
  scale_x_datetime(date_labels = "%d/%m", date_breaks = "24 hours", expand = c(0,0)) +
  theme(axis.ticks.x.bottom = element_blank(),
        axis.text.x.bottom = element_text(angle = 45, vjust = 0.5, hjust = 1, size = 17),
        axis.text.y.left = element_text(size = 17),
        text = element_text(family = "titillium", colour = "black"),
        axis.line.y = element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.major.y = element_line(colour = "lightgrey", linetype = "dotted", size = 0.3),
        axis.ticks = element_line(color = "lightgrey"),
        panel.background = element_blank(),
        plot.margin = margin(t=2,
                         r=10,
                         b=3,
                         l=5),
        plot.title = element_text(size = 50, face = "bold"),
        plot.subtitle = element_text(size = 35, face = "italic"),
        axis.title = element_text(size = 13),
        axis.title.x = element_text(margin = margin(t = 10), size = 25),
        axis.title.y = element_text(margin = margin(r = 10), size = 25)) +
  ylab("Proportion of hateful tweets") +
  xlab("Time when the tweet was sent") +
  labs(title = "The patterns of Spanish online hate speech",
       subtitle = "The % of hateful tweets increases the days after games, specially when Spain doesn't win.") +
  annotate(
    "text", x = as.POSIXct("2022-12-09 06:00"), y = 5.5,
    label = "Match",
    color = "black",
    size = 7
  ) +
  annotate(geom = "point", 
           x = as.POSIXct("2022-12-08 12:00"), y = 5.5, color = "grey", size = 3, shape = 16, alpha = 0.8) +
  annotate(
    "text", x = as.POSIXct("2022-12-09 06:00"), y = 5.1,
    label = "Win",
    color = "black",
    size = 7
  ) +
  annotate(geom = "point", 
           x = as.POSIXct("2022-12-08 12:00"), y = 5.1, color = "darkgreen", size = 3, shape = 16) +
annotate(
    "text", x = as.POSIXct("2022-12-09 06:00"), y = 4.8,
    label = "  Draw",
    color = "black",
    size = 7
  ) +
  annotate(geom = "point", 
           x = as.POSIXct("2022-12-08 12:00"), y = 4.8, color = "orange", size = 3, shape = 16) +
annotate(
    "text", x = as.POSIXct("2022-12-09 06:00"), y = 4.5,
    label = " Loss",
    color = "black",
    size = 7
  ) +
  annotate(geom = "point", 
           x = as.POSIXct("2022-12-08 12:00"), y = 4.5, color = "red", size = 3, shape = 16) +
  annotate(geom = "segment", x = as.POSIXct("2022-12-08 02:00"), xend = as.POSIXct("2022-12-09 22:00"), y = 5.3, yend = 5.3, color = "black", alpha = 0.5, size = 0.5) +
  
  annotation_custom(rasterGrob(germanyflag, width = unit(1, "npc"), height = unit(1, "npc")),
                           xmin = as.POSIXct("2022-11-26 06:00:00"), xmax = as.POSIXct("2022-11-27 06:00:00"),
                           ymin = 5.5, ymax = 5.9) + #Germany
  annotation_custom(rasterGrob(crflag, width = unit(1, "npc"), height = unit(1, "npc")),
                           xmin = as.POSIXct("2022-11-22 03:00:00"), xmax = as.POSIXct("2022-11-23 03:00:00"),
                           ymin = 5.5, ymax = 5.9) + #Costa Rica
  annotation_custom(rasterGrob(japanflag, width = unit(1, "npc"), height = unit(1, "npc")),
                           xmin = as.POSIXct("2022-11-30 03:00:00"), xmax = as.POSIXct("2022-12-01 03:00:00"),
                           ymin = 5.5, ymax = 5.9) + #Japan
  annotation_custom(rasterGrob(morflag, width = unit(1, "npc"), height = unit(1, "npc")),
                           xmin = as.POSIXct("2022-12-04 22:00:00"), xmax = as.POSIXct("2022-12-05 22:00:00"),
                           ymin = 5.5, ymax = 5.9) + #Morocco
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-26 06:00:00"),
           xend = as.POSIXct("2022-11-27 06:00:00"),
           y = 5.9, yend = 5.9,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-27 06:00:00"),
           xend = as.POSIXct("2022-11-27 06:00:00"),
           y = 5.9, yend = 5.5,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-26 06:00:00"),
           xend = as.POSIXct("2022-11-26 06:00:00"),
           y = 5.5, yend = 5.9,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-26 06:00:00"),
           xend = as.POSIXct("2022-11-27 06:00:00"),
           y = 5.5, yend = 5.5,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-22 03:00:00"),
           xend = as.POSIXct("2022-11-23 03:00:00"),
           y = 5.9, yend = 5.9,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-23 03:00:00"),
           xend = as.POSIXct("2022-11-23 03:00:00"),
           y = 5.9, yend = 5.5,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-22 03:00:00"),
           xend = as.POSIXct("2022-11-22 03:00:00"),
           y = 5.5, yend = 5.9,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-22 03:00:00"),
           xend = as.POSIXct("2022-11-23 03:00:00"),
           y = 5.5, yend = 5.5,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-30 03:00:00"),
           xend = as.POSIXct("2022-12-01 03:00:00"),
           y = 5.9, yend = 5.9,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-12-01 03:00:00"),
           xend = as.POSIXct("2022-12-01 03:00:00"),
           y = 5.9, yend = 5.5,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-30 03:00:00"),
           xend = as.POSIXct("2022-11-30 03:00:00"),
           y = 5.5, yend = 5.9,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-30 03:00:00"),
           xend = as.POSIXct("2022-12-01 03:00:00"),
           y = 5.5, yend = 5.5,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-12-04 22:00:00"),
           xend = as.POSIXct("2022-12-05 22:00:00"),
           y = 5.9, yend = 5.9,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-12-05 22:00:00"),
           xend = as.POSIXct("2022-12-05 22:00:00"),
           y = 5.9, yend = 5.5,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-12-04 22:00:00"),
           xend = as.POSIXct("2022-12-04 22:00:00"),
           y = 5.5, yend = 5.9,
           color = "black", size = 0.3) +
  annotate(geom = "segment",
           x = as.POSIXct("2022-12-04 22:00:00"),
           xend = as.POSIXct("2022-12-05 22:00:00"),
           y = 5.5, yend = 5.5,
           color = "black", size = 0.3) +
  # Annotation for Germany (diagonal segment)
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-27 06:00:00"),
           xend = as.POSIXct("2022-11-27 18:00:00"),
           y = 5.5, yend = 5.3,
           color = "black", size = 0.3) +
  # Annotation for Costa Rica (diagonal segment)
  annotate(geom = "segment",
           x = as.POSIXct("2022-11-23 03:00:00"),
           xend = as.POSIXct("2022-11-23 15:00:00"),
           y = 5.5, yend = 5.3,
           color = "black", size = 0.3) +
  # Annotation for Japan (diagonal segment)
  annotate(geom = "segment",
           x = as.POSIXct("2022-12-01 03:00:00"),
           xend = as.POSIXct("2022-12-01 18:00:00"),
           y = 5.5, yend = 5.3,
           color = "black", size = 0.3) +
  # Annotation for Morocco (diagonal segment)
  annotate(geom = "segment",
           x = as.POSIXct("2022-12-05 22:00:00"),
           xend = as.POSIXct("2022-12-06 14:00:00"),
           y = 5.5, yend = 5.3,
           color = "black", size = 0.3)


tmp_file <- tempfile(fileext = ".png")
ggsave(filename = tmp_file, width = 7, height = 4.5, dpi = 300, device = "png")
browseURL(tmp_file)

ggsave("general.jpeg", width = 7.5, height = 4.5, dpi = 300)

```


# Inside Spain vs. Morocco

```{r}
dfull2 <- dfull2 %>% 
  filter(created_at>="2022-12-06 14:00:00" & created_at<"2022-12-07 00:00:00") %>% 
  mutate(created_at = ymd_hms(created_at),
         created_at = format(created_at, "%Y-%m-%d %H:%M:00"),
         created_at = ymd_hms(created_at)) %>% 
  mutate(hate = factor(hate, levels = c("No", "Yes")),
         created_at = round_date(created_at, unit = "10 minutes")) %>% 
  group_by(created_at) %>% 
  summarise(No_count = sum(hate == "No", na.rm = TRUE),
            Yes_count = sum(hate == "Yes", na.rm = TRUE),
            Total = sum(No_count, Yes_count),
            Proportion = Yes_count/Total * 100)
```

```{r}
dfull2 %>% 
  ggplot() +
  aes(x = as.POSIXct(created_at), y = Proportion) +
  scale_x_datetime(date_labels = "%H:%M", date_breaks = "30 mins", expand = c(0,0)) +
  geom_rect(aes(xmin=as.POSIXct("2022-12-06 17:00"), xmax=as.POSIXct("2022-12-06 20:00"), ymin=(-Inf), ymax=Inf),
            alpha=0.05, show.legend=FALSE, fill = "lightgrey") +
  geom_line(size = 0.3) +
  annotate(geom = "segment", x = as.POSIXct("2022-12-06 18:35"), xend = as.POSIXct("2022-12-06 18:35"), y = 0, yend = Inf, alpha = 0.3, linetype = "dashed") +
  geom_smooth(method = "loess", size = 0.5) +
  scale_y_continuous(
    limits = c(0,6),
    labels = function(x) paste0(x, "%")) +
  theme(axis.text.x.bottom = element_text(angle = 45, vjust = 1, hjust = 1, size = 18),
        axis.text.y.left = element_text(size = 18),
        axis.ticks.x.bottom = element_blank(),
        text = element_text(family = "titillium", colour = "black"),
        axis.line.y = element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.major.y = element_line(colour = "lightgrey", linetype = "dotted", size = 0.3),
        axis.ticks = element_line(color = "lightgrey"),
        panel.background = element_blank(),
        plot.title = element_text(size = 50, face = "bold"),
        plot.subtitle = element_text(size = 35, face = "italic"),
        axis.title.y.left = element_text(size = 25),
        axis.title.x.bottom = element_text(vjust = 0, size = 25)) +
  ylab("Proportion of hateful tweets") +
  xlab("Time when the tweet was sent") +
  labs(title = "Hate speech increases before and after the match",
       subtitle = "When playing, the debate is less focused on hate compared to prior to and afterwards the game. ") +
  annotate(
    "text", x = as.POSIXct("2022-12-06 23:00"), y = 5.5,
    label = "Match",
    color = "black",
    size = 7
  ) +
  annotate(geom = "segment", x = as.POSIXct("2022-12-06 22:35"), xend = as.POSIXct("2022-12-06 22:45"), y = 5.2, yend = 5.2, linetype = "dashed", size = 0.2, alpha = 0.3) +
  annotate(
    "text", x = as.POSIXct("2022-12-06 23:05"), y = 5.2,
    label = "Overtime",
    color = "black",
    size = 7
  ) + 
  annotate(geom = "segment", x = as.POSIXct("2022-12-06 22:35"), xend = as.POSIXct("2022-12-06 23:20"), color = "black", alpha = 0.5, size = 0.35, y = 5.33, yend = 5.33) +
  annotate(geom = "point", 
           x = as.POSIXct("2022-12-06 22:40"), y = 5.5, color = "grey", size = 3, shape = 16, alpha = 0.8) +
  annotation_custom(rasterGrob(morflag, width = unit(1, "npc"), height = unit(1, "npc")),
                           xmin = as.POSIXct("2022-12-06 18:00:00"), xmax = as.POSIXct("2022-12-06 19:00:00"),
                           ymin = 5, ymax = 5.8) +
  annotate(geom = "segment",
         x = as.POSIXct("2022-12-06 18:00:00"),
         xend = as.POSIXct("2022-12-06 19:00:00"),
         y = 5, yend = 5,
         color = "black",
         size = 0.3) +
annotate(geom = "segment",
         x = as.POSIXct("2022-12-06 19:00:00"),
         xend = as.POSIXct("2022-12-06 19:00:00"),
         y = 5, yend = 5.8,
         color = "black",
         size = 0.3) +
annotate(geom = "segment",
         x = as.POSIXct("2022-12-06 19:00:00"),
         xend = as.POSIXct("2022-12-06 18:00:00"),
         y = 5.8, yend = 5.8,
         color = "black",
         size = 0.3) +
annotate(geom = "segment",
         x = as.POSIXct("2022-12-06 18:00:00"),
         xend = as.POSIXct("2022-12-06 18:00:00"),
         y = 5.8, yend = 5,
         color = "black",
         size = 0.3) 

tmp_file <- tempfile(fileext = ".png")
ggsave(filename = tmp_file, width = 7.5, height = 4.5, dpi = 300, device = "png")
browseURL(tmp_file)

ggsave("morocco.png", width = 7.5, height = 4.5, dpi = 300)
```


# T-tests

All: 

```{r}

s<-split(viz$Proportion,viz$match) #Splitting between moments with a match being played and moments with no match. 

#1st: Test for equal variances
var <- var.test(s$`0`, s$`1`, alternative = "two.sided")

#P-value is lower than 0.05, so we accept the H1: variances are different. 


#2nd step: Test for different means

tall <- t.test(s$`1`, s$`0` ,alternative="greater", mu=0, var.equal=FALSE, conf.level = 0.95)

#Not significant. Difference between both types of moments is not greater than 0. 


```

Morocco: 

```{r}

s<-split(viz$Proportion,viz$morocco)


var.test(s$`0`, s$`1`, alternative = "two.sided")

tmor <- t.test(s$`1`, s$`0`, alternative="greater",mu=0,var.equal=TRUE)

```


ANOVA: 

```{r}
viz <- viz %>% 
  mutate(match_which = case_when(costarica == 1 ~ "Costa Rica", 
                                 morocco == 1 ~ "Morocco", 
                                 germany == 1 ~ "Germany", 
                                 japan == 1 ~"Japan", 
                                 T ~ NA), 
         match_which = factor(match_which, levels = c("Costa Rica", "Morocco", "Germany", "Japan")))


tukey <- viz %>% rstatix::tukey_hsd(Proportion ~ match_which)
flextable::flextable(tukey) 

flextable::set_flextable_defaults(font.family = "Times New Roman", 
                                  text.align = "center",
                                  post_process_all = TRUE
                                  )

tukeytable <- tukey %>% 
  select(-c(term, null.value)) %>% 
  rename("Team 1" = "group1",
         "Team 2" = "group2", 
         "Estimate" = "estimate", 
         "CI Lower" = "conf.low", 
         "CI Higher" = "conf.high", 
         "P-value" = "p.adj", 
         "Significance" = "p.adj.signif") %>% 
  mutate(Estimate = round(Estimate, 4), 
         Significance = ifelse(Significance == "****", "***", ifelse(Significance == "ns", "-", ".")), 
         `P-value` = format(`P-value`, scientific = TRUE)) %>% 
  flextable::flextable() %>% 
  flextable::add_header_lines(values = "ANOVA Test for all comparisons between matches")
```




After the matches: 

```{r}
viz <- viz %>% 
  mutate(created_at = as.POSIXct(created_at),
    after = case_when(created_at >= "2022-11-23 22:00:00" & created_at <= "2022-11-24 22:00:00" ~ 1,
                      created_at >= "2022-11-28 00:00:00" & created_at <= "2022-11-29 00:00:00" ~ 1,
                      created_at >= "2022-12-02 00:00:00" & created_at <= "2022-12-03 00:00:00" ~ 1,
                      created_at >= "2022-12-06 20:00:00" & created_at <= "2022-12-07 20:00:00" ~ 1,
                           T ~ 0),
         after = factor(after, levels = c(0,1)))
```

```{r}
s<-split(viz$Proportion,viz$after)

var.test(s$`0`, s$`1`, alternative = "two.sided")

tafter <- t.test(s$`1`, s$`0`, alternative="greater",mu=0,var.equal=TRUE)
```

T-tests viz: 

```{r}
a <- broom::tidy(tall) %>% 
  mutate(Test = ".") %>% 
  select(c(Test, estimate, estimate1, estimate2, p.value, conf.low, conf.high, alternative))

b <- broom::tidy(tmor) %>% 
  mutate(Test = ".") %>% 
  select(c(Test, estimate, estimate1, estimate2, p.value, conf.low, conf.high, alternative))

c <- broom::tidy(tafter) %>% 
  mutate(Test = ".") %>% 
  select(c(Test, estimate, estimate1, estimate2, p.value, conf.low, conf.high, alternative))

ttests <-  rbind(a, b, c)

ttests$Test <- c("All matches", "Vs. Morocco", "After the matches")

teststable <- ttests %>% 
  rename("Mean of X" = "estimate1", 
         "Mean of Y" = "estimate2", 
         "P-value" = "p.value", 
         "Alternative" = "alternative") %>% 
  mutate(estimate = round(estimate, 4), 
         Significance = ifelse(`P-value` < 0.01, "***", ifelse(`P-value` < 0.05, "**", ifelse(`P-value` < 0.1, "*", ifelse(`P-value` >= 0.1, "-", "-")))), 
         `P-value` = ifelse(`P-value` < 0.1, format(`P-value`, scientific = T), round(`P-value`, 4))) %>% 
  select(-c(conf.low, conf.high)) %>% 
flextable::flextable() %>% 
  flextable::add_header_lines(values = "Comparisons in the level of hate speech (T-Tests)")

flextable::save_as_docx(teststable, tukeytable, path = "tables.docx", pr_section = prop_section(page_size(width = 12, height = 8)), page_mar(bottom = 1, top = 1, left=0, right = 0))
```
